move multi-device test into its own file
