+++ tests/api_test.py
@@ -1644,65 +1644,5 @@ class JaxprTest(jtu.JaxTestCase):
         """)
 
 
-class MultiDeviceTest(jtu.JaxTestCase):
-
-  def test_computation_follows_data(self):
-    if len(jax.devices()) < 2:
-      raise unittest.SkipTest("test requires multiple devices")
-
-    x = jax.device_put(1, jax.devices()[0])
-    y = jax.device_put(2, jax.devices()[0])
-    z = x + y
-    self.assertEqual(z, 3)
-    self.assertEqual(z.device_buffer.device(), jax.devices()[0])
-
-    x = jax.device_put(1, jax.devices()[1])
-    y = jax.device_put(2, jax.devices()[1])
-    z = x + y
-    self.assertEqual(z, 3)
-    self.assertEqual(z.device_buffer.device(), jax.devices()[1])
-
-    x = jax.device_put(1, jax.devices()[1])
-    y = 4
-    z = x + y
-    self.assertEqual(z, 5)
-    self.assertEqual(z.device_buffer.device(), jax.devices()[1])
-
-    x = jax.device_put(1, jax.devices()[0])
-    y = jax.device_put(2, jax.devices()[1])
-    self.assertRaisesRegex(
-        ValueError,
-        "primitive arguments must be colocated on the same device",
-        lambda: x + y)
-
-    x = jax.device_put(1, jax.devices()[1])
-    y = x.reshape((1, 1))
-    self.assertEqual(y.device_buffer.device(), jax.devices()[1])
-
-  def test_primitive_compilation_cache(self):
-    if len(jax.devices()) < 2:
-      raise unittest.SkipTest("test requires multiple devices")
-
-    primitive_computation = xla.primitive_computation
-    xla.xla_primitive_callable.cache_clear()  # clear op-by-op cache
-
-    count = [0]
-    def primitive_computation_and_count(*args, **kwargs):
-      count[0] += 1
-      return primitive_computation(*args, **kwargs)
-
-    x = jax.device_put(1, jax.devices()[1])
-
-    try:
-      xla.primitive_computation = primitive_computation_and_count
-      y = lax.add(x, x)
-      z = lax.add(y, y)
-    finally:
-      xla.primitive_computation = primitive_computation
-
-    self.assertEqual(count[0], 1)
-    self.assertEqual(y.device_buffer.device(), jax.devices()[1])
-    self.assertEqual(z.device_buffer.device(), jax.devices()[1])
-
 if __name__ == '__main__':
   absltest.main()
